MASSIVE Cursed monster v3
HUGE amounts of code added, all are trying to refactor previous blueprints into c++, and make the dialogue system fully self sufficient.

At this moment it is still incomplete, with much work to do when it comes to the widgets themselves and in general much more needs to be done concerning the interface.